move generate_page_summary
